Move <MediaStore> to its own file
