+++ MovieListFramework/Media Cache/LocalMediaLoader.swift
@@ -35,11 +35,3 @@ public final class LocalMediaLoader {
         }
     }
 }
-
-public protocol MediaStore {
-    typealias DeletionCompletion = (Error?) -> Void
-    typealias InsertionCompletion = (Error?) -> Void
-    
-    func deleteCachedMedia(completion: @escaping DeletionCompletion)
-    func insert(_ items: [MediaItem], timestamp: Date, completion: @escaping (InsertionCompletion))
-}
