+++ MovieListFramework/Media Cache/MediaStore.swift
@@ -1,0 +1,16 @@
+//
+//  MediaStore.swift
+//  MovieListFramework
+//
+//  Created by macbook on 13/09/2023.
+//
+
+import Foundation
+
+public protocol MediaStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsertionCompletion = (Error?) -> Void
+    
+    func deleteCachedMedia(completion: @escaping DeletionCompletion)
+    func insert(_ items: [MediaItem], timestamp: Date, completion: @escaping (InsertionCompletion))
+}
